backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
pypdf>=3.0.0
langchain>=0.1.11
langchain-google-genai
langchain-community
python-docx
python-pptx
langchain-text-splitters
langchain-ollama
google-generativeai
faiss-cpu==1.8.0
tiktoken==0.6.0
python-dotenv==1.0.1
langchain-huggingface
sentence-transformers
